requirements.txt
Django==2.2.6
djangorestframework
psycopg2-binary
gunicorn
django-heroku
whitenoise
dj-database-url
pyyaml
sqlalchemy
python-dotenv
data-wizard